Base Var + RPS modifications
Set up some of the basic variables that are necessary to the program as well as setting up the modifications to the RPS function for later in the program.

I need to start working on setting up the system that will be used for the rooms (i.e. use a funct, raw code, ect.)