Hospital Room 2
Added most of the exits for room 2 as well as the rooms description text for room 2. Added the description and implementation of the help command from the user.

Need to work on the next room as well as working on the 'grab' command so that the user can start interacting with the environment.