Formatting and Player Class
Moved around some variables so that the programming would be easier to set up. Added a player class that involves the player's inventory and current location. Get and set for cur_location. (numpy arrays can be added to tuples)

Need to set up the input commands for the player so that the code can be tested for formatting and errors.