Formatting and Calling Cards
Added some formatting so that the unformatted file is no the default file. Added the basic code for calling functions that are used to interact with the game (filled with 'pass')

Work on the help statement as I already have all of the stuff that I need.  Continue adding more functions for the calling cards.