Room Creation and Movement
Added the code for creating a basic room as well as two rooms. The movement system SHOULD all be working, all possibilities have NOT been tested. Description inputs for rooms are working, item pickups are not.

Need to add a system that allows for a room to store the items that can be picked up as well as ensuring that they can be stored in the player's inventory.