Description Texts
Added some of the description texts as well as some bug errors. Now shows an introduction message to the game as well as the allowed exits to a room.

Need to work on the display text of items that can be picked up within a room as well as items that can be used within a room.