Pickup Function
Made it so that the pickup function is fully functional as well as making it so that if there are items to pick up, their description text is displayed.

Need to work on making sure that the use function is fully functional as well as figuring out a way to ensure that the RPS code can be unlocked in one of the rooms.