Bug Fixing
Fixed a major bug for when the player is traveling between rooms and there are more than one room. Added a 3rd room where the coin item should be able to be used.

MUST make it so that the items can be used, as well as the setup for rock paper scissors.